update đọc excel nhân viên
2020.06.11.2

+++ src/DAO/NhanVienDAO.java
@@ -41,7 +41,7 @@ public class NhanVienDAO {
             String Sql ="SELECT * FROM tblnhanvien";
             ResultSet rs = st.executeQuery(Sql);
             while(rs.next()){
-                NhanVienDTO sp = new NhanVienDTO(rs.getString(1), rs.getString(2), rs.getString(3), rs.getDate(4), rs.getString(5), rs.getString(6),rs.getString(7), rs.getDouble(8),rs.getInt(9));
+                NhanVienDTO sp = new NhanVienDTO(rs.getString(1), rs.getString(2), rs.getString(3), rs.getDate(4), rs.getString(5), rs.getString(6),rs.getString(7), rs.getInt(8),rs.getInt(9));
                 list.add(sp);
             }
             conn.close();
